fix: resolve PR CI lint and message catalog coverage failures
Split editable workstation configuration state to satisfy Biome line limits,
register new guard message formatters in catalog coverage tests, and fix
inherited lint failures in factory graph and activity view model files.
